backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
    #   starlette
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
chardet==5.2.0
    # via reportlab
click==8.1.8
    # via uvicorn
distro==1.9.0
    # via openai
fastapi==0.115.11
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
jinja2==3.1.6
    # via -r requirements.in
jiter==0.9.0
    # via openai
markdown==3.7
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
openai==1.66.3
    # via -r requirements.in
pdfkit==1.0.0
    # via -r requirements.in
pillow==11.1.0
    # via reportlab
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
reportlab==4.3.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.38
    # via -r requirements.in
starlette==0.46.0
    # via fastapi
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.34.0
    # via -r requirements.in
email-validator==2.3.0
slowapi==0.1.9
